refactor(langgraph-python): organize schemas and clean up for beginners
File organization:
- Move agent schemas into apps/agent/src/a2ui/schemas/
  (flight_schema.json, booked_schema.json, dynamic_schema_prompt.md)
- Rename frontend streaming schema to streaming_flight_schema.json
- Rename booked-confirmation.json to booked_schema.json for consistency

Code clarity:
- Clean up ignoredTools with inline comments explaining each entry
- Modernize TypedDict imports (stdlib typing, not typing_extensions)
- Remove unused ChatOpenAI import from main.py
- Add section comments for import groups in main.py
