+++ examples/integrations/langgraph-python/apps/agent/main.py
@@ -5,10 +5,12 @@ It defines the workflow graph, state, tools, nodes and edges.
 
 from copilotkit import CopilotKitMiddleware
 from langchain.agents import create_agent
-from langchain_openai import ChatOpenAI
 
+# Data & state tools
 from src.query import query_data
 from src.todos import AgentState, todo_tools
+
+# A2UI tools (three approaches to agent-driven UI)
 from src.a2ui_fixed_schema import search_flights
 from src.a2ui_fixed_schema_streaming import search_flights_streaming
 from src.a2ui_dynamic_schema import generate_a2ui
@@ -19,18 +21,14 @@ agent = create_agent(
     middleware=[CopilotKitMiddleware()],
     state_schema=AgentState,
     system_prompt="""
-        You are a polished, professional demo assistant using CopilotKit and LangGraph. Only mention either when necessary.
+        You are a polished, professional demo assistant. Keep responses to 1-2 sentences.
 
-        Keep responses brief and polished — 1 to 2 sentences max. No verbose explanations.
-
-        When demonstrating charts, always call the query_data tool to fetch data first.
-        When asked to manage todos, enable app mode first, then manage todos.
-
-        When you see a log_a2ui_event tool result indicating a user action like "book_flight",
-        respond with a brief confirmation message. The UI updates optimistically on the frontend.
-
-        When asked to generate dynamic/custom UI, call generate_a2ui. It returns
-        A2UI operations directly — no need to call send_a2ui_json_to_client afterward.
+        Tool guidance:
+        - Charts: call query_data first, then render with the chart component.
+        - Todos: enable app mode first, then manage todos.
+        - A2UI actions: when you see a log_a2ui_event result (e.g. "book_flight"),
+          respond with a brief confirmation. The UI already updated on the frontend.
+        - Dynamic UI: call generate_a2ui directly. It handles rendering on its own.
     """,
 )
