+++ examples/integrations/langgraph-python/apps/agent/src/a2ui/schemas/booked_schema.json
