+++ examples/integrations/langgraph-python/apps/agent/src/a2ui/schemas/dynamic_schema_prompt.md
