+++ examples/integrations/langgraph-python/apps/agent/src/a2ui/schemas/flight_schema.json
