+++ examples/integrations/langgraph-python/apps/app/src/a2ui/booked_schema.json
