+++ examples/integrations/langgraph-python/apps/app/src/app/api/copilotkit/streaming_flight_schema.json
